Controllers refactrored
