+++ controllers/usersControllers.js
@@ -1,0 +1,114 @@
+const gravatar = require("gravatar");
+const bcrypt = require("bcryptjs");
+const jwt = require("jsonwebtoken");
+const keys = require("../config/keys");
+
+// Load Input Validation
+const validateRegisterInput = require("../validation/register");
+const validateLoginInput = require("../validation/login");
+
+// Loade user model
+const User = require("../models/User");
+
+module.exports = postUserRegisterController = (req, res, next) => {
+  const { errors, isValid } = validateRegisterInput(req.body);
+
+  // Check Validation
+  if (!isValid) {
+    return res.status(400).json(errors);
+  }
+
+  User.findOne({ email: req.body.email }).then(user => {
+    errors.email = "Email already exists";
+    if (user) {
+      return res.status(400).json(errors);
+    } else {
+      const avatar = gravatar.url(req.body.email, {
+        s: "200", //Size
+        r: "pg", //Rating
+        d: "mm" //Default
+      });
+
+      const newUser = new User({
+        name: req.body.name,
+        email: req.body.email,
+        avatar,
+        password: req.body.password,
+        isAdmin: false
+      });
+
+      bcrypt.genSalt(10, (err, salt) => {
+        bcrypt.hash(newUser.password, salt, (err, hash) => {
+          if (err) throw err;
+          newUser.password = hash;
+          newUser
+            .save()
+            .then(user => res.json(user))
+            .catch(err => console.log(err));
+        });
+      });
+    }
+  });
+};
+
+module.exports = postUserLoginController = (req, res, next) => {
+  const { errors, isValid } = validateLoginInput(req.body);
+
+  // Check Validation
+  if (!isValid) {
+    return res.status(400).json(errors);
+  }
+
+  const email = req.body.email;
+  const password = req.body.password;
+
+  //Find user by email
+  User.findOne({ email }).then(user => {
+    // Check for user
+    if (!user) {
+      errors.email = "User not found";
+      return res.status(404).json(errors);
+    }
+
+    // Check password
+    bcrypt.compare(password, user.password).then(isMatch => {
+      if (isMatch) {
+        // User Matched
+
+        const payload = {
+          id: user.id,
+          name: user.name,
+          avatar: user.avatar,
+          isAdmin: user.isAdmin
+        }; // crate JWT Payload
+
+        //jwt.sign(payload, secretOrPrivateKey, [options, callback])
+        // Sing Token
+        jwt.sign(
+          payload,
+          keys.secretOrKey,
+          { expiresIn: 3600 },
+          (err, token) => {
+            res.json({
+              success: true,
+              token: "Bearer " + token
+            });
+          }
+        );
+        if (user.isAdmin) {
+        }
+      } else {
+        errors.password = "Password incorrect";
+        return res.status(400).json({ errors });
+      }
+    });
+  });
+};
+
+module.exports = getCurrentUserController = (req, res, next) => {
+  res.json({
+    id: req.user.id,
+    name: req.user.name,
+    email: req.user.password
+  });
+};
