+++ models/Movie.js
@@ -8,11 +8,16 @@ const MovieSchema = new Schema({
     type: String,
     required: true
   },
-
   duration: {
     type: Number, // seconds
     required: true
-  }
+  },
+  shows: [
+    {
+      type: Schema.Types.ObjectId,
+      ref: "shows"
+    }
+  ]
 });
 
-module.exports = Movie = mongoose.model("movies", MovieSchema);
+module.exports = Movie = mongoose.model("movies", MovieSchema, "movies");
