+++ routes/api/users.js
@@ -1,17 +1,10 @@
 const express = require("express");
 const router = express.Router();
-const gravatar = require("gravatar");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
-const keys = require("../../config/keys");
 const passport = require("passport");
-
-// Load Input Validation
-const validateRegisterInput = require("../../validation/register");
-const validateLoginInput = require("../../validation/login");
-
-// Loade user model
-const User = require("../../models/User");
+// Load Controllers
+const postUserRegisterController = require("../../controllers/usersControllers");
+const postUserLoginController = require("../../controllers/usersControllers");
+const getCurrentUserController = require("../../controllers/usersControllers");
 
 // @route   GET api/users/test
 // @desc    Tests users route
@@ -21,104 +14,13 @@ router.get("/test", (req, res) => res.json({ msg: "Users Works" }));
 // @route   POST api/users/test
 // @desc    Tests users route
 // @access  Public
-router.post("/register", (req, res) => {
-  const { errors, isValid } = validateRegisterInput(req.body);
-
-  // Check Validation
-  if (!isValid) {
-    return res.status(400).json(errors);
-  }
-
-  User.findOne({ email: req.body.email }).then(user => {
-    errors.email = "Email already exists";
-    if (user) {
-      return res.status(400).json(errors);
-    } else {
-      const avatar = gravatar.url(req.body.email, {
-        s: "200", //Size
-        r: "pg", //Rating
-        d: "mm" //Default
-      });
-
-      const newUser = new User({
-        name: req.body.name,
-        email: req.body.email,
-        avatar,
-        password: req.body.password,
-        isAdmin: false
-      });
-
-      bcrypt.genSalt(10, (err, salt) => {
-        bcrypt.hash(newUser.password, salt, (err, hash) => {
-          if (err) throw err;
-          newUser.password = hash;
-          newUser
-            .save()
-            .then(user => res.json(user))
-            .catch(err => console.log(err));
-        });
-      });
-    }
-  });
-});
+router.post("/register", postUserRegisterController);
 
 // @route   POST api/users/login
 // @desc    Login User / Return JWT Token
 // @access  Public
 
-router.post("/login", (req, res) => {
-  const { errors, isValid } = validateLoginInput(req.body);
-
-  // Check Validation
-  if (!isValid) {
-    return res.status(400).json(errors);
-  }
-
-  const email = req.body.email;
-  const password = req.body.password;
-
-  //Find user by email
-  User.findOne({ email }).then(user => {
-    // Check for user
-    if (!user) {
-      errors.email = "User not found";
-      return res.status(404).json(errors);
-    }
-
-    // Check password
-    bcrypt.compare(password, user.password).then(isMatch => {
-      if (isMatch) {
-        // User Matched
-
-        const payload = {
-          id: user.id,
-          name: user.name,
-          avatar: user.avatar,
-          isAdmin: user.isAdmin
-        }; // crate JWT Payload
-
-        //jwt.sign(payload, secretOrPrivateKey, [options, callback])
-        // Sing Token
-        jwt.sign(
-          payload,
-          keys.secretOrKey,
-          { expiresIn: 3600 },
-          (err, token) => {
-            res.json({
-              success: true,
-              token: "Bearer " + token
-            });
-          }
-        );
-        if (user.isAdmin) {
-        }
-      } else {
-        errors.password = "Password incorrect";
-        return res.status(400).json({ errors });
-      }
-    });
-  });
-});
+router.post("/login", postUserLoginController);
 
 // @route   GET api/users/current
 // @desc    Return current user
@@ -126,12 +28,7 @@ router.post("/login", (req, res) => {
 router.get(
   "/current",
   passport.authenticate("jwt", { session: false }),
-  (req, res) => {
-    res.json({
-      id: req.user.id,
-      name: req.user.name,
-      email: req.user.password
-    });
-  }
+  getCurrentUserController
 );
+
 module.exports = router;
